feat(summary): hardcode all RusProfile data into AI customer analysis
Ensure all parsed RusProfile/DaData statistics appear in customer analysis
regardless of LLM behavior. Added _hardcode_profile_data() that overrides
LLM output with real data: name, INN, region, status, director, ОКВЭД,
financials, procurement history (as customer + supplier), arbitration,
ФССП, licenses. Frontend updated with new fields and fixed 0-value handling.
Also fixed rusprofile_id overflow (IntegerField → BigIntegerField).

+++ backend/apps/customers/models.py
@@ -35,7 +35,7 @@ class CustomerProfile(models.Model):
     risk_indicators = models.JSONField(default=dict)
 
     # Мета
-    rusprofile_id = models.IntegerField(null=True, blank=True)
+    rusprofile_id = models.BigIntegerField(null=True, blank=True)
     dadata_updated_at = models.DateTimeField(null=True, blank=True)
     rusprofile_updated_at = models.DateTimeField(null=True, blank=True)
     rusprofile_failed = models.BooleanField(default=False)
